requirements.txt
celery
redis

vastai-cli

torch
torchvision
diffusers

ffmpeg-python
opencv-python

boto3

safetensors[torch]
accelerate
s3transfer

tqdm

python-dotenv